+++ src/Features/Admin/Categories/EditCategory/index.js
@@ -23,7 +23,9 @@ const EditCategory = () => {
   ]);
 
   const saveHandler = () => {
-    validateOnSubmit(fields, setFields) ? setOnSaveSuccess(true) : setOnSaveSuccess(false);
+    const { validateArray, isValid } = validateOnSubmit(fields);
+    setFields(validateArray);
+    isValid ? setOnSaveSuccess(true) : setOnSaveSuccess(false);
   };
 
   const buttons = {
