+++ src/Features/Admin/OrderHistory/AddOrder/index.js
@@ -61,7 +61,9 @@ const AddOrder = () => {
   ]);
 
   const saveHandler = () => {
-    validateOnSubmit(fields, setFields) ? setOnSaveSuccess(true) : setOnSaveSuccess(false);
+    const { validateArray, isValid } = validateOnSubmit(fields);
+    setFields(validateArray);
+    isValid ? setOnSaveSuccess(true) : setOnSaveSuccess(false);
   };
 
   const buttons = {
