+++ src/Features/Admin/Users/AddUser/index.js
@@ -89,7 +89,9 @@ const AddUser = () => {
   ]);
 
   const saveHandler = () => {
-    validateOnSubmit(fields, setFields) ? setOnSaveSuccess(true) : setOnSaveSuccess(false);
+    const { validateArray, isValid } = validateOnSubmit(fields);
+    setFields(validateArray);
+    isValid ? setOnSaveSuccess(true) : setOnSaveSuccess(false);
   };
 
   const buttons = {
