+++ src/Features/Admin/Users/EditUser/index.js
@@ -88,8 +88,11 @@ const AddUser = () => {
   ]);
 
   const saveHandler = () => {
-    validateOnSubmit(fields, setFields) ? setOnSaveSuccess(true) : setOnSaveSuccess(false);
+    const { validateArray, isValid } = validateOnSubmit(fields);
+    setFields(validateArray);
+    isValid ? setOnSaveSuccess(true) : setOnSaveSuccess(false);
   };
+
   const buttons = {
     button: [
       {
