+++ src/Features/Admin/Vendors/AddVendor/index.js
@@ -114,7 +114,9 @@ const AddVendor = () => {
   ]);
 
   const saveHandler = () => {
-    validateOnSubmit(fields, setFields) ? setOnSaveSuccess(true) : setOnSaveSuccess(false);
+    const { validateArray, isValid } = validateOnSubmit(fields);
+    setFields(validateArray);
+    isValid ? setOnSaveSuccess(true) : setOnSaveSuccess(false);
   };
 
   const buttons = {
