+++ src/Features/Customer/Dashboard/index.js
@@ -4,8 +4,8 @@ import { Grid, Typography, Box } from '@material-ui/core';
 import { makeStyles } from '@material-ui/core/styles';
 import ArrowBackIcon from '@material-ui/icons/ArrowBack';
 
-import MainTab from '../../components/CardMenus/Tabs';
-import SideMenu from '../../components/sideMenu';
+import MainTab from '../../../components/CardMenus/Tabs';
+import SideMenu from '../../../components/sideMenu';
 
 function Dashboard() {
   const useStyles = makeStyles(() => ({
