+++ src/Features/Customer/Dashboard/route.js
@@ -1,5 +1,5 @@
-import { isPublicRoute } from '../../routes/Permission';
-import RouteNames from '../../routes/RouteNames';
+import { isPublicRoute } from '../../../routes/Permission';
+import RouteNames from '../../../routes/RouteNames';
 import Dashboard from './index';
 
 export const dashboardRoute = {
