+++ src/routes/RouteConfig.js
@@ -3,7 +3,7 @@ import { orderHistorydRoute, addOrderdRoute, editOrderdRoute } from '../Features
 import { addUserRoute, editUserRoute, usersRoute } from '../Features/Admin/Users/route';
 import { addVendorsRoute, editVendorsRoute, vendorsRoute } from '../Features/Admin/Vendors/route';
 import { loginRoute, signUpRoute, resetPasswordRoute, forgetPasswordRoute } from '../Features/Auth/route';
-import { dashboardRoute } from '../Features/Dashboard/route';
+import { dashboardRoute } from '../Features/Customer/Dashboard/route';
 import { homeRoute } from '../Features/Home';
 import { profileRoute } from '../Features/Profile';
